update : カードを表示できるようにした
