+++ src/app/students/courses/page.tsx
@@ -1,14 +1,19 @@
 "use client";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
+import { Carousel, CarouselContent, CarouselItem, CarouselNext, CarouselPrevious } from "@/components/ui/carousel";
+import { Progress } from "@/components/ui/progress";
 import { Separator } from "@/components/ui/separator";
 import { SidebarTrigger } from "@/components/ui/sidebar";
-import CardCarousel from "@/features/courses/components/CardCarousel";
-import type { Course } from "@/features/courses/types/CardCarousel";
+import type { SlideCard } from "@/features/courses/types/CardSlider";
+import Image from "next/image";
+import Link from "next/link";
 
 // ダミーデータ
 // ダミーデータのallCoursesとrelatedCoursesはidが重複してはいけません。
 // ここでは、allCoursesのidは1から10、relatedCoursesのidは11から20としています。
 // START:dummy
-const allCourses: Course[] = [];
+const allCourses: SlideCard[] = [];
 for (let i = 1; i <= 10; i++) {
   allCourses.push({
     id: i,
@@ -19,7 +24,7 @@ for (let i = 1; i <= 10; i++) {
   });
 }
 
-const relatedCourses: Course[] = [];
+const relatedCourses: SlideCard[] = [];
 for (let i = 11; i <= 20; i++) {
   relatedCourses.push({
     id: i,
@@ -50,11 +55,70 @@ export default function CoursesPage() {
         */}
         {/* Most Popular */}
         <h1 className="text-xl font-bold pl-6">Most Popular</h1>
-        <CardCarousel courses={allCourses} id={1} />
+        {/* カードスライダー */}
+        <div>
+          <Carousel
+            opts={{
+              align: "start",
+            }}
+            className="w-full max-w-sm"
+          >
+            <CarouselContent>
+              {allCourses.map((item, _) => (
+                <CarouselItem key={item.id}>
+                  <CardSlider {...allCourses[item.id]} />
+                </CarouselItem>
+              ))}
+            </CarouselContent>
+            <CarouselPrevious />
+            <CarouselNext />
+          </Carousel>
+        </div>
         {/* Related Courses */}
         <h1 className="text-xl font-bold pl-6">Related Courses</h1>
-        <CardCarousel courses={relatedCourses} id={2} />
+        {/* カードスライダー */}
       </div>
     </>
+  );
+}
+
+// カードコンポーネント
+function CardSlider({
+  id,
+  title,
+  description,
+  achievementLevel,
+  maxAchievementLevel,
+}: { id: number; title: string; description: string; achievementLevel: number; maxAchievementLevel: number }) {
+  const progressValue = (achievementLevel / maxAchievementLevel) * 100;
+
+  return (
+    <Card className="w-full max-w-[255px] shadow-md rounded-[36px] mb-11">
+      <div className="relative w-full h-[184px]">
+        <Image src={"/courseLogo.webp"} alt={title} layout="fill" objectFit="cover" className="rounded-t-[36px]" />
+      </div>
+
+      <CardHeader className="pl-3 pb-2 pt-2">
+        <CardTitle className="text-lg font-semibold leading-tight">{title}</CardTitle>
+        <CardDescription className="text-sm text-gray-500 leading-normal">{description}</CardDescription>
+      </CardHeader>
+
+      <CardContent className="flex flex-col gap-2">
+        {/* プログレスバー */}
+        <div className="w-full flex items-center justify-between">
+          <Progress value={progressValue} className="h-2 bg-gray-200 w-full" />
+          <p className="min-w-[50px] max-w-[70px] text-[9px] text-right text-gray-500 ml-2">
+            {achievementLevel} / {maxAchievementLevel} 完了
+          </p>
+        </div>
+
+        {/* ボタン */}
+        <div className="w-full flex justify-center mt-2">
+          <Link href={`/students/courses/${id}`} className="rounded-md shadow-md">
+            <Button className="w-[93px] h-[40px] bg-[#327fd6]">はじめる</Button>
+          </Link>
+        </div>
+      </CardContent>
+    </Card>
   );
 }
